Start boot2docker if it's available
This adds conditional tasks that we don't want to accidentally become
the default task, but need to be evaluated early. To avoid them
unintentionally becoming the default task, I've moved the `run` task to
the first position, forcing it to always be the default.

Also update the documentation to remove the boot2docker-centric steps
now they're automated, and add coloured error output functions in
Makefile.