Stop docker-compose building our image twice
The djangogulpserve container uses the same image as the django
container, so we should explicitly indicate that's what we want to
avoid building it twice. If djangogulpserve needed to be run on its own
this would break, but we never actually need that because it only
exists for serving assets during development.

Also make the the environment variable passed to `--port` consistent
with the one used for `--host`.