Add a hacky test-runner
Our Dockerfile only installs production requirements, and
docker-compose doesn't install development requirements till it runs
the container, so to run the tests outside the container we need to
install the dependencies.

To avoid getting lots of distracting noise when running tests, we're
running the `pip install` quietly, but it will still error if
something's amiss.