Indent conditionals in Makefile
Just because it's make, that doesn't mean that conditional blocks
shouldn't be indented for clarity.

One slightly odd thing: because we're using the default recipe prefix
character of \t, indented recipe lines *start* with that and then have
the same number of indent spaces following it. This seemed the least
objectionable layout of the lines (spaces followed by a \t? ew.).