Upgrade (hackily) to node 4.x
We were pinning to node 0.10.x for no particular reason. The PPA we
were using [is now
deprecated](https://chrislea.com/2014/07/09/joining-forces-nodesource/)
in favour of [NodeSource's
distributions](https://github.com/nodesource/distributions/).

We [shouldn't really be piping to
bash](https://www.seancassidy.me/dont-pipe-to-your-shell.html), of
course, but I'm not going to fix that right now.