Install node modules outside this directory
To avoid trashing node modules that're installed as part of building
the docker container, we need to install them to a location outside
`/app`. The best way to do this is to install them as global modules,
using the `--global` flag.

For node/npm to treat global modules as local, we have to set the
`$NODE_PATH` to include the global install location
(`/usr/lib/node_modules`). Because of the way node installs dependencies,
though, we also need to include the `node_modules` directory within our
module's installed directory in the `$NODE_PATH`.

Then we have to ensure that Bower installs to the right place
(taking care to use `cwd` *and* `directory`, to avoid Bower's installer
just appending the `directory` value to the location from which the
command is run). Finally, we have to ensure that all the places that
refer to the bower modules know to look in the new locations.

We [shouldn't be setting `cwd` in our
`.bowerrc`](https://github.com/bower/spec/blob/master/config.md#cwd),
but will fix that in a later commit by explicitly passing `cwd` to
commands and passing around appropriate environment variables. Because
of this, this commit may not work outside the Docker environment.