Compile assets outside of this directory
To avoid docker-compose mounting this directory into the container
overriding all the assets that have been built during the Docker build,
we should compile all the assets to a location on the container that's
outside the current working directory.

This, as with 2b3a7b7, is a hack that probably won't work very well
outside the containers and should be made configurable with liberal
application of environment variables.