Add confirmation complete page and handle error states
Eror states can be when the disbursement is no longer valid, either
because its resolution has changed or because the details (prisoner
location, amount) are not valid in NOMIS. If this occurs the payment
is reset to pending and the user is returned to the payment detail
page, where hopefully they will be informed what went wrong.

Theoretically a payment could get stuck in 'preconfirmed' whilst having
invalid details, which at the moment would need to be manually resolved
(by us).