Update tests to work with Django 2.2
In Django 2.2 the post method changed to not allow `None` to be passed
as form data, to fix the tests we are making the test use the same value
as the button in `new_credits.html`.

We tried removing the key value pair altogether, but that meant the form
was not valid and tests failed. While we could not find the key being
directly used anywhere, there's some confusing code where the key
`submit_new` (or `submit_manual`) is used to select the form needed
in the views[2]. And it seems the value is only used in the JS[3,4] to
determine if a confirmation dialog box is displayed.

1.https://github.com/ministryofjustice/money-to-prisoners-cashbook/blob/master/mtp_cashbook/templates/cashbook/new_credits.html
2. https://github.com/ministryofjustice/money-to-prisoners-cashbook/blob/master/mtp_cashbook/apps/cashbook/views.py#L107-L108
3. https://github.com/ministryofjustice/money-to-prisoners-cashbook/blob/master/mtp_cashbook/assets-src/javascripts/modules/batch-validation.js#L61
4. https://github.com/ministryofjustice/money-to-prisoners-cashbook/blob/master/mtp_cashbook/assets-src/javascripts/modules/confirm-manual.js#L27-L28