Fix issue with `choices` override being ignored
There are 2 forms where the `choices` of the form
field is overriden dynamically to set it to the list
of credits coming from the API.

These worked fine but in Django 5.0+ it's causing
problems with some tests failing because of submitted
credit being "not a valid choice".

After much digging the issue seems to be that the
second element of the 2-tuple must be a string.

The cached property used to populate the choices
however is a list where the 2nd element of the
2-tuple is a list itself and Django doesn't like it.

Interesting this is not listed as a breaking change
in the Django 5.0 release notes - presumably not
many Django applications do something like this or
have the 2nd element as a non-string.

NOTE: The `credit_choices` cached properties remain
untouched as the 2nd field of the 2-tuple is used
in other parts of the application.